# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Бэкенд «Умного города» — обращения жителей МКД к управляющей компании, плановые работы, справочники домов/УК/аварийных служб. Один инстанс = одна УК.\n\nРоли: representative (представитель УК), dispatcher (диспетчер), resident (житель). Роль определяется самим пользователем на бэкенде — в запросах отдельно не передаётся.",
        "title": "MAXITO API",
        "contact": {
            "name": "MAXITO"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/representative/assignments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-assignments"
                ],
                "summary": "Все назначения диспетчер-дом",
                "operationId": "representativeListAssignments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.DispatcherHouse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/representative/dispatchers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-dispatchers"
                ],
                "summary": "Список диспетчеров",
                "operationId": "representativeListDispatchers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-dispatchers"
                ],
                "summary": "Добавить диспетчера",
                "operationId": "representativeCreateDispatcher",
                "parameters": [
                    {
                        "description": "ФИО и телефон",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_representative.CreateDispatcherRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/representative/dispatchers/csv": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обязательные колонки: full_name, phone.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-dispatchers"
                ],
                "summary": "Массово добавить диспетчеров из CSV",
                "operationId": "representativeImportDispatchersCSV",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV-файл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_services.ImportReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/representative/dispatchers/{dispatcher_id}/houses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-assignments"
                ],
                "summary": "Дома конкретного диспетчера",
                "operationId": "representativeListDispatcherHouses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID диспетчера",
                        "name": "dispatcher_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.House"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-assignments"
                ],
                "summary": "Назначить диспетчеру дома",
                "operationId": "representativeAssignHouses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID диспетчера",
                        "name": "dispatcher_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Список ID домов",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_representative.AssignHousesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/maxito_internal_services.ImportReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/representative/dispatchers/{dispatcher_id}/houses/{house_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "representative-assignments"
                ],
                "summary": "Снять с диспетчера дом",
                "operationId": "representativeUnassignHouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID диспетчера",
                        "name": "dispatcher_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID дома",
                        "name": "house_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "нет содержимого"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "назначение не найдено",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/representative/houses/unassigned": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "representative-assignments"
                ],
                "summary": "Дома без закреплённого диспетчера",
                "operationId": "representativeListUnassignedHouses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/maxito_internal_models.House"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apidoc.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код ошибки; присутствует только у части\nэндпоинтов (см. описание конкретного эндпоинта), у остальных отсутствует.",
                    "type": "string",
                    "example": "bad_request"
                },
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "internal_handlers_representative.AssignHousesRequest": {
            "type": "object",
            "required": [
                "house_ids"
            ],
            "properties": {
                "house_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "internal_handlers_representative.CreateDispatcherRequest": {
            "type": "object",
            "required": [
                "full_name",
                "phone"
            ],
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.Appeal": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealAttachment"
                    }
                },
                "author": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "author_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discovered_at": {
                    "description": "когда житель заметил проблему (необязательно)",
                    "type": "string"
                },
                "entrance_number": {
                    "description": "null = весь дом",
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "importance": {
                    "$ref": "#/definitions/maxito_internal_models.Importance"
                },
                "problem_type": {
                    "$ref": "#/definitions/maxito_internal_models.ProblemType"
                },
                "problem_type_id": {
                    "type": "integer"
                },
                "reason": {
                    "$ref": "#/definitions/maxito_internal_models.Reason"
                },
                "reason_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/maxito_internal_models.AppealStatus"
                },
                "status_changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealStatusChange"
                    }
                },
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.AppealSubscription"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "wants_recalculation": {
                    "type": "boolean"
                }
            }
        },
        "maxito_internal_models.AppealAttachment": {
            "type": "object",
            "properties": {
                "appeal_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_models.DispatcherHouse": {
            "type": "object",
            "properties": {
                "assigned_at": {
                    "type": "string"
                },
                "assigned_by": {
                    "description": "Representative who assigned",
                    "type": "integer"
                },
                "assigner": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "dispatcher": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "dispatcher_id": {
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_models.House": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "appeals": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Appeal"
                    }
                },
                "chat_invite_link": {
                    "type": "string"
                },
                "construction_year": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "dispatchers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.DispatcherHouse"
                    }
                },
                "entrances_count": {
                    "type": "integer"
                },
                "floors_count": {
                    "description": "необязательно",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Notification"
                    }
                },
                "number": {
                    "type": "string"
                },
                "residents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.Resident"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "maxito_internal_models.Importance": {
            "type": "string",
            "enum": [
                "normal",
                "important"
            ],
            "x-enum-varnames": [
                "ImportanceNormal",
                "ImportanceImportant"
            ]
        },
...
        "maxito_internal_models.Resident": {
            "type": "object",
            "properties": {
                "apartment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "entrance_number": {
                    "type": "integer"
                },
                "house": {
                    "$ref": "#/definitions/maxito_internal_models.House"
                },
                "house_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/maxito_internal_models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "maxito_internal_models.Role": {
            "type": "string",
            "enum": [
                "representative",
                "dispatcher",
                "resident"
            ],
            "x-enum-varnames": [
                "RoleRepresentative",
                "RoleDispatcher",
                "RoleResident"
            ]
        },
        "maxito_internal_models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dispatcher_houses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_models.DispatcherHouse"
                    }
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_user_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resident_profile": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/maxito_internal_models.Resident"
                        }
                    ]
                },
                "role": {
                    "$ref": "#/definitions/maxito_internal_models.Role"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "maxito_internal_services.ImportReport": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/maxito_internal_services.RowResult"
                    }
                },
                "skipped": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "maxito_internal_services.RowResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                },
                "status": {
                    "description": "created | skipped | error",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT, полученный через POST /auth/max. Передавать как \"Bearer \u0026lt;token\u0026gt;\". Для локальной отладки без MAX бэкенд также принимает X-Max-User-Id (+X-Max-User-Phone при первой привязке), если на сервере включён ALLOW_DEV_HEADERS.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalKey": {
            "description": "Общий секрет между сервером бота и бэкендом — только для /internal/*.",
            "type": "apiKey",
            "name": "X-Internal-Key",
            "in": "header"
        }
    }
}